pyproject.toml
[tool.poetry]
name = "kg-rag"
version = "0.1.0"
description = ""
authors = ["Teilo Millet"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
llama-index = "^0.10.39"
llama-index-llms-openai = "^0.1.21"
llama-index-llms-groq = "^0.1.4"
llama-index-llms-mistralai = "^0.1.12"
python-dotenv = "^1.0.1"
llama-index-graph-stores-kuzu = "^0.1.4"
fire = "^0.6.0"
llama-index-embeddings-openai = "^0.1.10"
llama-index-embeddings-mistralai = "^0.1.4"
loguru = "^0.7.2"
neo4j = "^5.20.0"
llama-index-graph-stores-neo4j = "^0.1.4"
langchain = "^0.2.1"
langchain-community = "^0.2.1"
langchain-groq = "^0.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
